autoconf: Fix build with external libjpeg on FreeBSD.
The build would use the bundled libjpeg regardless of the setting of the
`--with-libjpeg` configure switch because the incorrect make variable
would be set.

Also the incorrect make variables for setting the related LDFLAGS and
CFLAGS were used, so these would not be picked up by the build.

This patch fixes both issues.